requirements.txt
asgiref==3.6.0
boto3==1.26.89
botocore==1.29.89
cffi==1.15.1
cryptography==39.0.2
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
jmespath==1.0.1
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
whitenoise==6.4.0
